+++ src/EventExport.cpp
@@ -8,11 +8,8 @@
 #include <ll/api/utils/ErrorUtils.h>
 
 #define GET_INSTANCE_MACRO(EXPORT_NAME, CLASS_NAME)                                                                    \
-    RemoteCall::exportAs("iListenAttentively", EXPORT_NAME, [](CompoundTag* info) -> CLASS_NAME* {                     \
-        if (info && ll::event::isEventSerializedObj(*info) && (*info)["_pointer_"].getId() == Tag::Type::Int64) {      \
-            return (CLASS_NAME*)((uintptr_t)((*info)["_pointer_"].get<Int64Tag>()));                                   \
-        }                                                                                                              \
-        return nullptr;                                                                                                \
+    RemoteCall::exportAs("iListenAttentively", EXPORT_NAME, [](uintptr_t info) -> CLASS_NAME* {                        \
+        return reinterpret_cast<CLASS_NAME*>(info);                                                                    \
     });
 
 namespace ila {