+++ src/event/LseEvent.cpp
@@ -8,7 +8,7 @@ namespace ila::event {
 void LseEvent::serialize(CompoundTag& nbt) const {
     Cancellable::serialize(nbt);
     if (!getData()) return;
-    for (auto& [key, value] : *getData()) {
+    for (auto& [key, value] : getData()->mTags) {
         if (key == "eventId") continue;
         nbt[key] = value;
     }
@@ -17,7 +17,7 @@ void LseEvent::serialize(CompoundTag& nbt) const {
 void LseEvent::deserialize(CompoundTag const& nbt) {
     Cancellable::deserialize(nbt);
     if (!getData()) return;
-    for (auto& [key, value] : nbt) {
+    for (auto& [key, value] : nbt.mTags) {
         if (key == "eventId") continue;
         (*getData())[key] = value;
     }