+++ xmake.lua
@@ -1,12 +1,10 @@
 add_rules("mode.debug", "mode.release")
 
 add_repositories("liteldev-repo https://github.com/LiteLDev/xmake-repo.git")
-add_repositories("miracleforest-repo https://github.com/MiracleForest/xmake-repo.git")
 
-add_requires("levilamina 1.0.1", {configs = {target_type = "server"}})
+add_requires("levilamina 1.1.0-rc.1", {configs = {target_type = "server"}})
 add_requires("levibuildscript 0.3.0")
 add_requires("legacyremotecall 0.9.0-rc.1")
-add_requires("ilistenattentively 0.2.3")
 
 if not has_config("vs_runtime") then
     set_runtimes("MD")
@@ -35,8 +33,7 @@ target("iListenAttentively-LseExport")
     )
     add_packages(
         "levilamina",
-        "legacyremotecall",
-        "ilistenattentively"
+        "legacyremotecall"
     )
     set_exceptions("none")
     set_kind("shared")